抽离完成！
